+++ src/collective/contact/contactlist/browser/excelexport.py
@@ -1,3 +1,4 @@
+from collective.contact.contactlist.api import get_contacts
 from zope.component import adapts
 from zope.i18n import translate
 
@@ -16,7 +17,7 @@ class ContactListDataSource(BaseContentsDataSource):
         return "%s.xls" % (self.context.getId())
 
     def get_objects(self):
-        return [r.to_object for r in self.context.contacts if r.to_object]
+        return list(get_contacts(self.context))
 
 
 class ExportLinkViewlet(ViewletBase):